Attribute Endpoints

+++ Solomonlol.EcommerseApi/Services/AttributeService.cs
@@ -7,20 +7,24 @@ using Solomonlol.EcommerseApi.MyResults;
 
 namespace Solomonlol.EcommerseApi.Services
 {
-    public class AtrributeService : IAttributeService, IAttributeValueService
+    public class AttributeService : IAttributeService, IAttributeValueService
     {
         private readonly IMapper _mapper;
         private readonly ApplicationContext _db;
-        public AtrributeService(IMapper mapper, ApplicationContext db)
+        public AttributeService(IMapper mapper, ApplicationContext db)
         {
             _db = db;
             _mapper = mapper;
         }
-        public async Task<Result> AddAttribute(ProductAttributeDto item, CancellationToken ct = default)
+        public async Task<Result> AddAttribute(string categoryName, ProductAttributeDto item, CancellationToken ct = default)
         {
-            var checkCategory = await _db.Categories.FindAsync(item.CategoryId, ct);
+            var checkCategory = await _db.Categories.FirstOrDefaultAsync(c => c.Name == categoryName, ct);
             if (checkCategory != null)
             {
+                var exists = await _db.ProductAttributes.AnyAsync(a => a.Name == item.Name, ct);
+
+                if (exists) return Result.Failure("Attribute with this name already exists");
+
                 var attribute = _mapper.Map<ProductAttribute>(item);
                 await _db.ProductAttributes.AddAsync(attribute, ct);
                 return await _db.SaveChangesAsync(ct) > 0
@@ -29,21 +33,26 @@ namespace Solomonlol.EcommerseApi.Services
             }
             else return Result.Failure("Category was not found");
         }
-        public async Task<Result> DeleteAttribute(string name, CancellationToken ct = default)
+        public async Task<Result> DeleteAttribute(string categoryName, string attributeName, CancellationToken ct = default)
         {
-            var checkAttribute = await _db.ProductAttributes.FirstOrDefaultAsync(x => x.Name == name, ct);
+            var checkCategory = await _db.Categories.FirstOrDefaultAsync(c => c.Name == categoryName, ct);
+
+            if (checkCategory == null) 
+                return Result.Failure("Category was not found");
+
+            var checkAttribute = await _db.ProductAttributes.FirstOrDefaultAsync(x => x.Name == attributeName, ct);
             if (checkAttribute != null)
             {
                 _db.ProductAttributes.Remove(checkAttribute);
                 return await _db.SaveChangesAsync(ct) > 0 
-                    ? Result.Success(name) 
+                    ? Result.Success(attributeName) 
                     : Result.Failure("Cannot save changes to database");
             }
             else return Result.Failure("Attribute was not found");
         }
-        public async Task<Result> UpdateAttribute(ProductAttributeDto item, CancellationToken ct = default)
+        public async Task<Result> UpdateAttribute(string categoryName, ProductAttributeDto item, CancellationToken ct = default)
         {
-            var checkCategory = await _db.Categories.FindAsync(item.CategoryId, ct);
+            var checkCategory = await _db.Categories.FirstOrDefaultAsync(c => c.Name == categoryName, ct);
             if (checkCategory != null)
             {
                 var attribute = await _db.ProductAttributes.FindAsync(item.Id, ct);
@@ -59,8 +68,13 @@ namespace Solomonlol.EcommerseApi.Services
             else return Result.Failure("Category was not found.");
         }
 
-        public async Task<Result> AddAttributeValue(ProductAttributeValueDto item, CancellationToken ct = default)
+        public async Task<Result> AddAttributeValue(string productName, ProductAttributeValueDto item, CancellationToken ct = default)
         {
+            var productCheck = await _db.Products.FirstOrDefaultAsync(p => p.Name == productName, ct);
+
+            if (productCheck == null)
+                return Result.Failure($"Product with name '{productName}' was not found.");
+
             var valueCheck = await _db.ProductAttributeValues
                 .FirstOrDefaultAsync(a=>
                 a.ProductAttributeId==item.ProductAttributeId && 
@@ -77,12 +91,22 @@ namespace Solomonlol.EcommerseApi.Services
         }
 
 
-        public async Task<Result> DeleteAttributeValue(int productId, int productAttributeId, CancellationToken ct = default)
+        public async Task<Result> DeleteAttributeValue(string productName, string productAttributeName, CancellationToken ct = default)
         {
+            var productCheck = await _db.Products.FirstOrDefaultAsync(p => p.Name == productName, ct);
+
+            if (productCheck == null)
+                return Result.Failure($"Product with name '{productName}' was not found.");
+
+            var productAttributeCheck = await _db.ProductAttributes.FirstOrDefaultAsync(a => a.Name == productAttributeName, ct);
+
+            if (productAttributeCheck == null)
+                return Result.Failure($"Attribute with name '{productAttributeName}' in product '{productName}' was not found.");
+
             var valueCheck = await _db.ProductAttributeValues
                 .FirstOrDefaultAsync(a =>
-                a.ProductAttributeId == productId &&
-                a.ProductId == productAttributeId, ct);
+                a.ProductAttribute.Name == productAttributeName &&
+                a.Product.Name == productName, ct);
             if (valueCheck != null)
             {
                 _db.ProductAttributeValues.Remove(valueCheck);
@@ -93,9 +117,18 @@ namespace Solomonlol.EcommerseApi.Services
             else return Result.Failure("Attribute value was not found.");
         }
 
-
-        public async Task<Result> UpdateAttributeValue(ProductAttributeValueDto item, CancellationToken ct = default)
+        public async Task<Result> UpdateAttributeValue(string productName, string productAttributeName, ProductAttributeValueDto item, CancellationToken ct = default)
         {
+            var productCheck = await _db.Products.FirstOrDefaultAsync(p => p.Name == productName, ct);
+
+            if (productCheck == null)
+                return Result.Failure($"Product with name '{productName}' was not found.");
+
+            var productAttributeCheck = await _db.ProductAttributes.FirstOrDefaultAsync(a => a.Name == productAttributeName, ct);
+
+            if (productAttributeCheck == null)
+                return Result.Failure($"Attribute with name '{productAttributeName}' in product '{productName}' was not found.");
+
             var valueCheck = await _db.ProductAttributeValues
                 .FirstOrDefaultAsync(a =>
                 a.ProductAttributeId == item.ProductAttributeId &&
